fix: convert `configure.js` to ESM
