+++ test/configure/getPlatformPackage.test.mjs
@@ -1,7 +1,7 @@
 // @ts-check
 import { deepEqual, equal, throws } from "node:assert/strict";
 import { describe, it } from "node:test";
-import { getPlatformPackage } from "../../scripts/configure.js";
+import { getPlatformPackage } from "../../scripts/configure.mjs";
 import { spy } from "../spy.mjs";
 
 describe("getPlatformPackage()", () => {
