+++ test/configure/mockParams.mjs
@@ -3,8 +3,8 @@
 
 /**
  * Returns mock parameters.
- * @param {Partial<import("../../scripts/configure.js").ConfigureParams>} [overrides]
- * @returns {import("../../scripts/configure.js").ConfigureParams}
+ * @param {Partial<import("../../scripts/configure.mjs").ConfigureParams>} [overrides]
+ * @returns {import("../../scripts/configure.mjs").ConfigureParams}
  */
 export function mockParams(overrides) {
   return {
