+++ test/pack.test.mjs
@@ -174,12 +174,13 @@ describe("npm pack", () => {
       "scripts/config-plugins/plugins/withIosBaseMods.mjs",
       "scripts/config-plugins/provider.mjs",
       "scripts/config-plugins/types.mjs",
-      "scripts/configure.js",
+      "scripts/configure-projects.js",
+      "scripts/configure.mjs",
       "scripts/helpers.js",
       "scripts/init.mjs",
       "scripts/parseargs.js",
       "scripts/schema.js",
-      "scripts/template.js",
+      "scripts/template.mjs",
       "scripts/validate-manifest.js",
       "test-app.gradle",
       "test_app.rb",
